# docs/api/openapi.yaml
# 이 파일은 scripts/export_openapi.py 로 자동 생성됩니다.
# 직접 수정하지 말고 backend/app 의 모델을 고친 뒤 다시 생성해 주세요.
openapi: 3.1.0
info:
  title: Gildam API
  description: 길담 MVP API입니다. 검증된 코스 DB(schema v3.1)를 원천으로 조건 필터링·귀가 가능성 검증·설명 가능한 추천을 제공합니다.
  version: 0.3.0
paths:
  /health:
    get:
      summary: Health Check
      description: 서버 상태와 함께 데이터 정합성 요약을 돌려줍니다.
      operationId: health_check_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Health Check Health Get
  /api/meta/conditions:
    get:
      summary: Read Condition Options
      description: 조건 선택지의 단일 정의. 프론트엔드 하드코딩 드리프트를 막습니다.
      operationId: read_condition_options_api_meta_conditions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Read Condition Options Api Meta Conditions Get
  /api/courses/{course_id}:
    get:
      summary: Read Course Detail
      operationId: read_course_detail_api_courses__course_id__get
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: string
          title: Course Id
      - name: duration
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Duration
      - name: simulate
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Simulate
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/recommendations:
    post:
      summary: Read Recommendations
      operationId: read_recommendations_api_recommendations_post
      parameters:
      - name: simulate
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Simulate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    CourseDestination:
      properties:
        name:
          type: string
          title: Name
        latitude:
          type: number
          title: Latitude
        longitude:
          type: number
          title: Longitude
      type: object
      required:
      - name
      - latitude
      - longitude
      title: CourseDestination
    CourseDetailResponse:
      properties:
        schemaVersion:
          type: string
          title: Schemaversion
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        region:
          type: string
          title: Region
        courseType:
          type: string
          title: Coursetype
        departurePoint:
          type: string
          title: Departurepoint
        departurePointName:
          type: string
          title: Departurepointname
        applicableDays:
          items:
            type: string
          type: array
          title: Applicabledays
        timeType:
          type: string
          title: Timetype
        verificationStatus:
          type: string
          enum:
          - VERIFIED
          - VERIFIED_MAP
          - PARTIALLY_VERIFIED
          - NEEDS_RECHECK
          - TEMPORARILY_UNAVAILABLE
          - REFERENCE
          - BLOCKED
          title: Verificationstatus
        publishable:
          type: boolean
          title: Publishable
        exposureTier:
          type: string
          enum:
          - PUBLIC
          - MANUAL_REVIEW
          - DEMO_ONLY
          - BLOCKED
          title: Exposuretier
        verifiedDate:
          type: string
          title: Verifieddate
        thumbnailUrl:
          type: string
          title: Thumbnailurl
        tags:
          items:
            type: string
          type: array
          title: Tags
        preferences:
          items:
            type: string
          type: array
          title: Preferences
        fatigueLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          title: Fatiguelevel
        fatigueScore:
          type: number
          title: Fatiguescore
        fatigueExplanation:
          $ref: '#/components/schemas/FatigueExplanation'
        durationMinutes:
          type: integer
          title: Durationminutes
        totalMinutes:
          $ref: '#/components/schemas/MinutesRange'
        walkingMinutes:
          type: integer
          title: Walkingminutes
        walkingMinutesRange:
          $ref: '#/components/schemas/MinutesRange'
        insideWalkingMinutes:
          type: integer
          title: Insidewalkingminutes
          default: 0
        transferCount:
          type: integer
          title: Transfercount
        roundTripTransitMinutes:
          type: integer
          title: Roundtriptransitminutes
        localTransitMinutes:
          type: integer
          title: Localtransitminutes
          default: 0
        stayMinutes:
          type: integer
          title: Stayminutes
          default: 0
        recommendationReasons:
          items:
            type: string
          type: array
          title: Recommendationreasons
        description:
          type: string
          title: Description
        cautions:
          items:
            type: string
          type: array
          title: Cautions
        manualChecks:
          items:
            type: string
          type: array
          title: Manualchecks
        itinerary:
          items:
            $ref: '#/components/schemas/ItineraryItem'
          type: array
          title: Itinerary
        localFood:
          items:
            $ref: '#/components/schemas/LocalFood'
          type: array
          title: Localfood
        localPoints:
          items:
            $ref: '#/components/schemas/LocalPoint'
          type: array
          title: Localpoints
        scenePrompts:
          items:
            type: string
          type: array
          title: Sceneprompts
        sources:
          items:
            $ref: '#/components/schemas/DataSource'
          type: array
          title: Sources
        primaryDestination:
          $ref: '#/components/schemas/CourseDestination'
        mapUrl:
          type: string
          title: Mapurl
        directionsUrl:
          type: string
          title: Directionsurl
        kakaoMapUrl:
          type: string
          title: Kakaomapurl
        kakaoDirectionsUrl:
          type: string
          title: Kakaodirectionsurl
        routeLinks:
          items:
            $ref: '#/components/schemas/RouteLink'
          type: array
          title: Routelinks
        returnFeasibility:
          $ref: '#/components/schemas/ReturnFeasibilityModel'
        exposureNotice:
          anyOf:
          - $ref: '#/components/schemas/ExposureNotice'
          - type: 'null'
        dataSnapshotDate:
          type: string
          title: Datasnapshotdate
        dataSourceName:
          type: string
          title: Datasourcename
        dataSourceUrl:
          type: string
          title: Datasourceurl
      type: object
      required:
      - schemaVersion
      - id
      - title
      - region
      - courseType
      - departurePoint
      - departurePointName
      - applicableDays
      - timeType
      - verificationStatus
      - publishable
      - exposureTier
      - verifiedDate
      - thumbnailUrl
      - tags
      - preferences
      - fatigueLevel
      - fatigueScore
      - fatigueExplanation
      - durationMinutes
      - totalMinutes
      - walkingMinutes
      - walkingMinutesRange
      - transferCount
      - roundTripTransitMinutes
      - recommendationReasons
      - description
      - itinerary
      - primaryDestination
      - mapUrl
      - directionsUrl
      - kakaoMapUrl
      - kakaoDirectionsUrl
      - returnFeasibility
      - dataSnapshotDate
      - dataSourceName
      - dataSourceUrl
      title: CourseDetailResponse
    CourseRecommendationSummary:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        region:
          type: string
          title: Region
        thumbnailUrl:
          type: string
          title: Thumbnailurl
        tags:
          items:
            type: string
          type: array
          title: Tags
        courseType:
          type: string
          title: Coursetype
        fatigueLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          title: Fatiguelevel
        fatigueScore:
          type: number
          title: Fatiguescore
        fatigueExplanation:
          $ref: '#/components/schemas/FatigueExplanation'
        durationMinutes:
          type: integer
          title: Durationminutes
        durationMinMinutes:
          type: integer
          title: Durationminminutes
        durationMaxMinutes:
          type: integer
          title: Durationmaxminutes
        walkingMinutes:
          type: integer
          title: Walkingminutes
        transferCount:
          type: integer
          title: Transfercount
        roundTripTransitMinutes:
          type: integer
          title: Roundtriptransitminutes
        verificationStatus:
          type: string
          title: Verificationstatus
        exposureTier:
          type: string
          title: Exposuretier
        recommendationReasons:
          items:
            type: string
          type: array
          title: Recommendationreasons
        recommendationScore:
          type: number
          title: Recommendationscore
        scoreBreakdown:
          $ref: '#/components/schemas/RecommendationScoreBreakdown'
        returnFeasibility:
          $ref: '#/components/schemas/ReturnFeasibilityModel'
      type: object
      required:
      - id
      - title
      - region
      - thumbnailUrl
      - tags
      - courseType
      - fatigueLevel
      - fatigueScore
      - fatigueExplanation
      - durationMinutes
      - durationMinMinutes
      - durationMaxMinutes
      - walkingMinutes
      - transferCount
      - roundTripTransitMinutes
      - verificationStatus
      - exposureTier
      - recommendationReasons
      - recommendationScore
      - scoreBreakdown
      - returnFeasibility
      title: CourseRecommendationSummary
    DataSource:
      properties:
        label:
          type: string
          title: Label
        url:
          type: string
          title: Url
        checkedDate:
          type: string
          title: Checkeddate
        verificationStatus:
          type: string
          title: Verificationstatus
      type: object
      required:
      - label
      - url
      - checkedDate
      - verificationStatus
      title: DataSource
    ExcludedCourse:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        reasons:
          items:
            $ref: '#/components/schemas/ExclusionReason'
          type: array
          title: Reasons
      type: object
      required:
      - id
      - title
      - reasons
      title: ExcludedCourse
    ExclusionReason:
      properties:
        code:
          type: string
          enum:
          - UNSUPPORTED_DEPARTURE
          - DAY_NOT_SUPPORTED
          - TIME_LIMIT_EXCEEDED
          - RETURN_NOT_FEASIBLE
          - MOBILITY_LIMIT_EXCEEDED
          - PREFERENCE_MISMATCH
          - BLOCKED_BY_EXPOSURE_POLICY
          - SCHEMA_INVALID
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
      - code
      - message
      title: ExclusionReason
    ExposureNotice:
      properties:
        tier:
          type: string
          enum:
          - PUBLIC
          - MANUAL_REVIEW
          - DEMO_ONLY
          - BLOCKED
          title: Tier
        title:
          type: string
          title: Title
        message:
          type: string
          title: Message
      type: object
      required:
      - tier
      - title
      - message
      title: ExposureNotice
    FatigueExplanation:
      properties:
        level:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          title: Level
        score:
          type: number
          title: Score
        calculatedLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          title: Calculatedlevel
        sourceLevel:
          anyOf:
          - type: string
            enum:
            - LOW
            - MEDIUM
            - HIGH
          - type: 'null'
          title: Sourcelevel
        resolution:
          type: string
          title: Resolution
        factors:
          items:
            $ref: '#/components/schemas/FatigueFactorModel'
          type: array
          title: Factors
        formula:
          type: string
          title: Formula
      type: object
      required:
      - level
      - score
      - calculatedLevel
      - resolution
      - factors
      - formula
      title: FatigueExplanation
    FatigueFactorModel:
      properties:
        key:
          type: string
          title: Key
        label:
          type: string
          title: Label
        value:
          type: integer
          title: Value
        unit:
          type: string
          title: Unit
        level:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          title: Level
        levelScore:
          type: integer
          title: Levelscore
        weight:
          type: number
          title: Weight
        contribution:
          type: number
          title: Contribution
        threshold:
          type: string
          title: Threshold
      type: object
      required:
      - key
      - label
      - value
      - unit
      - level
      - levelScore
      - weight
      - contribution
      - threshold
      title: FatigueFactorModel
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ItineraryItem:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        type:
          type: string
          enum:
          - transport
          - walk
          - place
          - food
          title: Type
        time:
          anyOf:
          - type: string
          - type: 'null'
          title: Time
        durationMinutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Durationminutes
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        segmentType:
          anyOf:
          - type: string
          - type: 'null'
          title: Segmenttype
        verificationStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Verificationstatus
        isTransfer:
          type: boolean
          title: Istransfer
          default: false
        place:
          anyOf:
          - $ref: '#/components/schemas/CourseDestination'
          - type: 'null'
        mapUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Mapurl
      type: object
      required:
      - id
      - name
      - type
      title: ItineraryItem
    LocalFood:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        tags:
          items:
            type: string
          type: array
          title: Tags
      type: object
      required:
      - id
      - name
      - description
      title: LocalFood
    LocalPoint:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        tags:
          items:
            type: string
          type: array
          title: Tags
      type: object
      required:
      - id
      - title
      - description
      title: LocalPoint
    MinutesRange:
      properties:
        min:
          type: integer
          title: Min
        plan:
          type: integer
          title: Plan
        max:
          type: integer
          title: Max
      type: object
      required:
      - min
      - plan
      - max
      title: MinutesRange
    MobilityScoreFactor:
      properties:
        score:
          type: number
          title: Score
        weight:
          type: number
          title: Weight
        weightedScore:
          type: number
          title: Weightedscore
        explanation:
          type: string
          title: Explanation
        fatigueScore:
          type: number
          title: Fatiguescore
        fatigueLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          title: Fatiguelevel
        walkingMinutes:
          type: integer
          title: Walkingminutes
        transferCount:
          type: integer
          title: Transfercount
      type: object
      required:
      - score
      - weight
      - weightedScore
      - explanation
      - fatigueScore
      - fatigueLevel
      - walkingMinutes
      - transferCount
      title: MobilityScoreFactor
    PreferenceMatchScoreFactor:
      properties:
        score:
          type: number
          title: Score
        weight:
          type: number
          title: Weight
        weightedScore:
          type: number
          title: Weightedscore
        explanation:
          type: string
          title: Explanation
        matchedCount:
          type: integer
          title: Matchedcount
        selectedCount:
          type: integer
          title: Selectedcount
        matchedPreferences:
          items:
            type: string
            enum:
            - NATURE_WALK
            - HISTORY_CULTURE
            - FOOD_MARKET
            - MEMORY
          type: array
          title: Matchedpreferences
      type: object
      required:
      - score
      - weight
      - weightedScore
      - explanation
      - matchedCount
      - selectedCount
      - matchedPreferences
      title: PreferenceMatchScoreFactor
    RecommendationMeta:
      properties:
        exposureMode:
          type: string
          title: Exposuremode
        evaluatedCount:
          type: integer
          title: Evaluatedcount
        blockedCount:
          type: integer
          title: Blockedcount
        schemaInvalidCount:
          type: integer
          title: Schemainvalidcount
        dataSnapshotDate:
          type: string
          title: Datasnapshotdate
        appliedMobility:
          type: string
          enum:
          - MIN_TRANSFER
          - LOW_BURDEN
          - ANY
          title: Appliedmobility
      type: object
      required:
      - exposureMode
      - evaluatedCount
      - blockedCount
      - schemaInvalidCount
      - dataSnapshotDate
      - appliedMobility
      title: RecommendationMeta
    RecommendationRequest:
      properties:
        departure:
          type: string
          enum:
          - GWANGJU_SONGJEONG
          - USQUARE
          title: Departure
        duration:
          type: string
          enum:
          - SIX_HOURS
          - FULL_DAY
          title: Duration
        preferences:
          items:
            type: string
            enum:
            - NATURE_WALK
            - HISTORY_CULTURE
            - FOOD_MARKET
            - MEMORY
          type: array
          minItems: 1
          title: Preferences
        mobility:
          type: string
          enum:
          - MIN_TRANSFER
          - LOW_BURDEN
          - ANY
          title: Mobility
          default: ANY
      type: object
      required:
      - departure
      - duration
      - preferences
      title: RecommendationRequest
    RecommendationResponse:
      properties:
        courses:
          items:
            $ref: '#/components/schemas/CourseRecommendationSummary'
          type: array
          title: Courses
        exclusions:
          items:
            $ref: '#/components/schemas/ExcludedCourse'
          type: array
          title: Exclusions
        suggestions:
          items:
            $ref: '#/components/schemas/RecommendationSuggestion'
          type: array
          title: Suggestions
        meta:
          $ref: '#/components/schemas/RecommendationMeta'
      type: object
      required:
      - courses
      - exclusions
      - meta
      title: RecommendationResponse
    RecommendationScoreBreakdown:
      properties:
        preferenceMatch:
          $ref: '#/components/schemas/PreferenceMatchScoreFactor'
        mobility:
          $ref: '#/components/schemas/MobilityScoreFactor'
        returnMargin:
          $ref: '#/components/schemas/ReturnMarginScoreFactor'
        localResource:
          $ref: '#/components/schemas/RecommendationScoreFactor'
        recordFit:
          $ref: '#/components/schemas/RecommendationScoreFactor'
      type: object
      required:
      - preferenceMatch
      - mobility
      - returnMargin
      - localResource
      - recordFit
      title: RecommendationScoreBreakdown
    RecommendationScoreFactor:
      properties:
        score:
          type: number
          title: Score
        weight:
          type: number
          title: Weight
        weightedScore:
          type: number
          title: Weightedscore
        explanation:
          type: string
          title: Explanation
      type: object
      required:
      - score
      - weight
      - weightedScore
      - explanation
      title: RecommendationScoreFactor
    RecommendationSuggestion:
      properties:
        code:
          type: string
          enum:
          - RELAX_DURATION
          - RELAX_MOBILITY
          - ADD_PREFERENCE
          - CHANGE_DEPARTURE
          - NO_ALTERNATIVE
          title: Code
        message:
          type: string
          title: Message
        availableCount:
          type: integer
          title: Availablecount
          default: 0
      type: object
      required:
      - code
      - message
      title: RecommendationSuggestion
    ReturnFeasibilityModel:
      properties:
        status:
          type: string
          enum:
          - FEASIBLE
          - TIGHT
          - NOT_FEASIBLE
          title: Status
        confidence:
          type: string
          enum:
          - CONFIRMED
          - NEEDS_DAY_OF_CHECK
          - UNVERIFIED
          title: Confidence
        departureTime:
          type: string
          title: Departuretime
        plannedReturnTime:
          type: string
          title: Plannedreturntime
        latestReturnTime:
          type: string
          title: Latestreturntime
        plannedTotalMinutes:
          type: integer
          title: Plannedtotalminutes
        worstCaseTotalMinutes:
          type: integer
          title: Worstcasetotalminutes
        allowedMinutes:
          type: integer
          title: Allowedminutes
        slackMinutes:
          type: integer
          title: Slackminutes
        lastActivityEndTime:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastactivityendtime
        lastReturnDeparture:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastreturndeparture
        lastReturnSlackMinutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lastreturnslackminutes
        bookingRequired:
          type: boolean
          title: Bookingrequired
          default: false
        messages:
          items:
            type: string
          type: array
          title: Messages
      type: object
      required:
      - status
      - confidence
      - departureTime
      - plannedReturnTime
      - latestReturnTime
      - plannedTotalMinutes
      - worstCaseTotalMinutes
      - allowedMinutes
      - slackMinutes
      title: ReturnFeasibilityModel
    ReturnMarginScoreFactor:
      properties:
        score:
          type: number
          title: Score
        weight:
          type: number
          title: Weight
        weightedScore:
          type: number
          title: Weightedscore
        explanation:
          type: string
          title: Explanation
        slackMinutes:
          type: integer
          title: Slackminutes
        status:
          type: string
          title: Status
      type: object
      required:
      - score
      - weight
      - weightedScore
      - explanation
      - slackMinutes
      - status
      title: ReturnMarginScoreFactor
    RouteLink:
      properties:
        order:
          type: integer
          title: Order
        fromName:
          type: string
          title: Fromname
        toName:
          type: string
          title: Toname
        transitUrl:
          type: string
          title: Transiturl
        walkUrl:
          type: string
          title: Walkurl
      type: object
      required:
      - order
      - fromName
      - toName
      - transitUrl
      - walkUrl
      title: RouteLink
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError